Update README.md
The Online Chat Web Application is a web-based platform designed to facilitate real-time communication and collaboration among users in the form of chat rooms. This platform aims to provide a user-friendly, interactive, and feature-rich environment for individuals and groups to engage in discussions, share information, and build communities. It serves as a versatile and dynamic communication tool, offering both public and private chat rooms to cater to various user needs